cloud_storage_functions/v1_product_catalog_inventory_sales_data/requirements.txt
pytz
Faker
Flask
google-cloud-storage